refactor: #275 Update dto und service.Closes #275

+++ backend/java/parser_service/src/main/java/com/voriq/parser_service/domain/dto/EngineDto.java
@@ -14,5 +14,5 @@ public class EngineDto {
     private FuelTypeDto fuelType;
     private String seriesCode;
     private String engineCode;
-    private Integer displacementCC;
+    private Double displacementCC;
 }
